+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to Rust's notion of
 - `note`:
   - `RandomSeed` and its functions `from_bytes` and `as_bytes` now public
   -  `Note::from_parts` now public
+  - added pub fn `random_seed`
 - `keys::Diversifier::from_bytes` now public
 
 ## [0.2.0] - 2022-06-24